feat: Refactor filter model structure and update documentation
- Moved the `Filter` model to a new `models` submodule for better organization within the `utils` package.
- Updated import paths across the codebase to reflect the new structure.
- Enhanced documentation for the `Filter` model, detailing its purpose and functionality in constructing SQLAlchemy `WHERE` clauses.
- Adjusted related tests to ensure compatibility with the refactored model location.
