+++ backend/src/reader/modules/routers/models/base/metadata/pagination_with_filters_query.py
@@ -5,7 +5,7 @@ __all__ = ("PaginationWithFiltersQuery",)
 
 from pydantic import Field, Json
 
-from reader.utils.filter import Filter
+from reader.utils.models.filter import Filter
 
 from .pagination_query import PaginationQuery
