+++ backend/src/reader/utils/__init__.py
@@ -2,6 +2,6 @@
 
 __all__ = ("Filter", "Singleton", "mount_static_files")
 
-from .filter import Filter
+from .models import Filter
 from .mount_static_files import mount_static_files
 from .singleton import Singleton
