+++ backend/src/reader/utils/models/__init__.py
@@ -1,0 +1,5 @@
+"""Utility models for the Reader application."""
+
+__all__ = ("Filter",)
+
+from .filter import Filter
